Fix parser to handle decimal scores (e.g., 3.5, -1.5)
- Update regex to capture decimal numbers: -?\d+\.?\d*
- Change score type from int to float in EvaluationResult
- Fixes 43+ parsing failures where judge intended decimal scores

+++ cot_quality_metrics/src/cot_quality_metrics/schemas.py
@@ -42,7 +42,7 @@ class EvaluationResult:
     """Result from a single rubric evaluation."""
 
     dimension: str
-    score: int
+    score: float  # Can be decimal (e.g., 3.5) or integer
     evidence: list[str]
     reasoning: str
